revno: 18210 message:   asha activity and beneficiary service recording
committer: Abyot Asalefew Gizaw <abyota@gmail.com>
branch nick: asha
timestamp: Fri 2015-04-03 16:39:43 +0200

+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/trackedentity/startup/TableAlteror.java
@@ -316,13 +316,32 @@ public class TableAlteror
 
         updateProgramExpressionUid();
         
-        executeSql( "DROP TABLE programstage_programindicators" );
+        if ( tableExists( "programstage_programindicators" ) )
+        {
+        	executeSql( "DROP TABLE programstage_programindicators" );
+            
+            log.info( "Dropped table programstage_ programindicators" );
+        }
+        
     }
 
     // -------------------------------------------------------------------------
     // Supporting methods
     // -------------------------------------------------------------------------
 
+    private boolean tableExists( String table )
+    {
+        try
+        {
+            statementManager.getHolder().queryForInteger( "select 1 from " + table );
+            return true;
+        }
+        catch ( Exception ex )
+        {
+            return false;
+        }
+    }
+    
     private void updateAggregateQueryBuilder()
     {
         StatementHolder holder = statementManager.getHolder();
